Consolidate time functions in library
