+++ Source/Lib/Common/Codec/EbUtility.c
@@ -5,6 +5,7 @@
 
 #include "EbDefinitions.h"
 #include "EbUtility.h"
+#include "EbSvtAv1Time.h"
 
 #ifdef _WIN32
 //#if  (WIN_ENCODER_TIMING || WIN_DECODER_TIMING)
@@ -550,7 +551,7 @@ const MiniGopStats* get_mini_gop_stats(const uint32_t mini_gop_index)
 }
 
 
-void EbStartTime(uint64_t *Startseconds, uint64_t *Startuseconds) {
+EB_API void EbStartTime(uint64_t *Startseconds, uint64_t *Startuseconds) {
 
 #if defined(__linux__) || defined(__APPLE__) //(LINUX_ENCODER_TIMING || LINUX_DECODER_TIMING)
     struct timeval start;
@@ -567,7 +568,7 @@ void EbStartTime(uint64_t *Startseconds, uint64_t *Startuseconds) {
 
 }
 
-void EbFinishTime(uint64_t *Finishseconds, uint64_t *Finishuseconds) {
+EB_API void EbFinishTime(uint64_t *Finishseconds, uint64_t *Finishuseconds) {
 
 #if defined(__linux__) || defined(__APPLE__) //(LINUX_ENCODER_TIMING || LINUX_DECODER_TIMING)
     struct timeval finish;
@@ -583,7 +584,7 @@ void EbFinishTime(uint64_t *Finishseconds, uint64_t *Finishuseconds) {
 #endif
 
 }
-void ComputeOverallElapsedTime(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration)
+EB_API void EbComputeOverallElapsedTime(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration)
 {
 #if defined(__linux__) || defined(__APPLE__) //(LINUX_ENCODER_TIMING || LINUX_DECODER_TIMING)
     long   mtime, seconds, useconds;
@@ -607,7 +608,7 @@ void ComputeOverallElapsedTime(uint64_t Startseconds, uint64_t Startuseconds, ui
 #endif
 
 }
-void EbComputeOverallElapsedTimeMs(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration)
+EB_API void EbComputeOverallElapsedTimeMs(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration)
 {
 #if defined(__linux__) || defined(__APPLE__) //(LINUX_ENCODER_TIMING || LINUX_DECODER_TIMING)
     long   mtime, seconds, useconds;
@@ -632,6 +633,78 @@ void EbComputeOverallElapsedTimeMs(uint64_t Startseconds, uint64_t Startuseconds
 
 }
 
+EB_API void EbSleep(uint64_t milliSeconds)
+{
+    if(milliSeconds) {
+#if defined(__linux__) || defined(__APPLE__)
+        struct timespec req,rem;
+        req.tv_sec=(int32_t)(milliSeconds/1000);
+        milliSeconds -= req.tv_sec * 1000;
+        req.tv_nsec = milliSeconds * 1000000UL;
+        nanosleep(&req,&rem);
+#elif _WIN32
+        Sleep((DWORD) milliSeconds);
+#else
+#error OS Not supported
+#endif
+    }
+}
+
+EB_API void EbInjector(uint64_t processed_frame_count,
+                       uint32_t injector_frame_rate)
+{
+#if defined(__linux__) || defined(__APPLE__)
+    uint64_t                  currentTimesSeconds = 0;
+    uint64_t                  currentTimesuSeconds = 0;
+    static uint64_t           startTimesSeconds;
+    static uint64_t           startTimesuSeconds;
+#elif _WIN32
+    static LARGE_INTEGER    startCount;               // this is the start time
+    static LARGE_INTEGER    counterFreq;              // performance counter frequency
+    LARGE_INTEGER           nowCount;                 // this is the current time
+#else
+#error OS Not supported
+#endif
+
+    double                 injectorInterval  = (double)(1<<16)/(double)injector_frame_rate;     // 1.0 / injector frame rate (in this case, 1.0/encodRate)
+    double                  elapsedTime;
+    double                  predictedTime;
+    int32_t                     bufferFrames = 1;         // How far ahead of time should we let it get
+    int32_t                     milliSecAhead;
+    static int32_t              firstTime = 0;
+
+    if (firstTime == 0)
+    {
+        firstTime = 1;
+
+#if defined(__linux__) || defined(__APPLE__)
+        EbStartTime((uint64_t*)&startTimesSeconds, (uint64_t*)&startTimesuSeconds);
+#elif _WIN32
+        QueryPerformanceFrequency(&counterFreq);
+        QueryPerformanceCounter(&startCount);
+#endif
+    }
+    else
+    {
+
+#if defined(__linux__) || defined(__APPLE__)
+        EbFinishTime((uint64_t*)&currentTimesSeconds, (uint64_t*)&currentTimesuSeconds);
+        EbComputeOverallElapsedTime(startTimesSeconds, startTimesuSeconds, currentTimesSeconds, currentTimesuSeconds, &elapsedTime);
+#elif _WIN32
+        QueryPerformanceCounter(&nowCount);
+        elapsedTime = (double)(nowCount.QuadPart - startCount.QuadPart) / (double)counterFreq.QuadPart;
+#endif
+
+        predictedTime = (processed_frame_count - bufferFrames) * injectorInterval;
+        milliSecAhead = (int32_t)(1000 * (predictedTime - elapsedTime ));
+        if (milliSecAhead>0)
+        {
+            //  timeBeginPeriod(1);
+            EbSleep(milliSecAhead);
+            //  timeEndPeriod (1);
+        }
+    }
+}
 
 uint32_t ns_quarter_off_mult[9/*Up to 9 part*/][2/*x+y*/][4/*Up to 4 ns blocks per part*/] =
 {
